+++ ksupol/lab_03/draw.h
@@ -5,15 +5,15 @@
 
 void set_colour(QColor *c, int bgc_col, int col);
 
-void draw_library(float sX, float sY, float eX, float eY, QColor *c, QGraphicsScene *scene);
+void draw_library(int sX, int sY, int eX, int eY, QColor *c, QGraphicsScene *scene);
 
-void draw_dda(float sX, float sY, float eX, float eY, QColor *cl, QGraphicsScene *scene);
+void draw_dda(int sX, int sY, int eX, int eY, QColor *cl, QGraphicsScene *scene);
 
-void draw_real(float sX, float sY, float eX, float eY, QColor *c, QGraphicsScene *scene);
+void draw_real(int sX, int sY, int eX, int eY, QColor *c, QGraphicsScene *scene);
 
-void draw_int(float sX, float sY, float eX, float eY, QColor *c, QGraphicsScene *scene);
+void draw_int(int sX, int sY, int eX, int eY, QColor *c, QGraphicsScene *scene);
 
-void draw_step(float sX, float sY, float eX, float eY, QColor *c, QGraphicsScene *scene);
+void draw_step(int sX, int sY, int eX, int eY, QColor *c, QGraphicsScene *scene);
 
 void draw_spectrum(QColor *c, int alg, float degr, QGraphicsScene *scene);
